Moved old files into new project structure
