+++ src/index.html
@@ -1,12 +1,48 @@
 <!DOCTYPE html>
-<html lang="en">
+<html>
 <head>
-    <meta charset="UTF-8">
-    <title>Title</title>
+<meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+<title>Adv Dupe Inspector - Subnet ROOT Labs</title>
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
+<link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.2.0/css/solid.css" integrity="sha384-wnAC7ln+XN0UKdcPvJvtqIH3jOjs9pnKnq9qX68ImXvOGz2JuFoEiCjT8jyZQX2z" crossorigin="anonymous">
+<link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.2.0/css/fontawesome.css" integrity="sha384-HbmWTHay9psM8qyzEKPc8odH4DsOuzdejtnr+OFtDmOcIVnhgReQ4GZBH7uwcjf6" crossorigin="anonymous">
 </head>
 <body>
 
-Hey! <b>The project</b> works!
+<div id="header">
+	<h1>Adv Dupe Inspector</h1>
+	<p>
+		See what's in your dupes. Select a dupe from your Gmod data folder (%STEAMDIR%/steamapps/common/GarrysMod/garrysmod/data/adv_duplicator) <b>OR</b> drag a file into this window to get started.
+		<br>
+		<a href="http://www.subnetroot.com">Powered by Subnet ROOT</a>
+	</p>
+	<div id="fileupload">
+		<i class="fas fa-file-upload"></i> Choose a dupe file
+		<div id="filestatus" class="hidden">
+			<i class="fas fa-cog fa-spin status-icon loading-icon"></i>
+			<i class="fas fa-check-circle status-icon accepted-icon"></i>
+			<i class="fas fa-times-circle status-icon rejected-icon"></i>
+			<i class="fas fa-exclamation-circle status-icon warning-icon"></i>
+			<span class="label"></span>
+		</div>
+		<input type="file">
+	</div>
+</div>
+
+<div id="treeview-container">
+	<div id="treeview">
+		<h2>Treeview</h2>
+		<details open>
+			<summary>(no file loaded)</summary>
+		</details>
+	</div>
+</div>
+
+<div id="inspector-container">
+	<div id="inspector">
+		<h2>Inspector area</h2>
+	</div>
+</div>
 
 </body>
 </html>
