+++ src/index.ts
@@ -1,10 +1,3 @@
 import './style.css';
 
-class TestMe {
-    doIt(): void {
-        console.log("hello world");
-    }
-}
-
-const tester = new TestMe();
-tester.doIt();
+import './app/main.js';
